Update README.md
Hi, I'm doing the project code review4. Congrats for completing the final! The project's objective and purpose are clearly stated in the file, and the app itself has a lovely user interface and provides nice experience. There're a few suggestions, 1. it would be more readable to add the new released feature from this version. 2. it is recommended in class that it's better to put typography and color palette and explain why in readme.